requirements.txt
selenium
webdriver-manager
pandas
lxml
python-dotenv
prettytable